+++ tests/components/businessInfo.react.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallowToJson } from 'enzyme-to-json';
 
-import { BusinessInfo } from './../client/src/components/BusinessInfo.jsx';
+import { BusinessInfo } from './../../client/src/components/BusinessInfo.jsx';
 
 describe('<BusinessInfo/>', () => {
 	let component = shallow(<BusinessInfo business={{}}/>);
