--- tests/test.js
@@ -1,3 +1,0 @@
-test('adds 1 + 2 to equal 3', () => {
-  expect(1+2).toBe(3);
-});
